+++ Installer/OS/kali23_1.py
@@ -304,7 +304,10 @@ python3 -m pip install tensorflow_cpu --break-system-packages
 export PYTHONPATH='""" + fissure_directory + """':$PYTHONPATH
 python3 ./fissure/generate_certificates.py
 ########## Verify ##########
-ls '""" + fissure_directory + """/certificates'
+test -f '""" + fissure_directory + """/certificates/server/server.key' &&
+test -f '""" + fissure_directory + """/certificates/server/server.key_secret' &&
+test -f '""" + fissure_directory + """/certificates/clients/client_0.key' &&
+test -f '""" + fissure_directory + """/certificates/clients/client_0.key_secret'
 """,True,'Minimum Install'))
 
 # Auto-Launch Sensor Node